[change] Move client/cli.go -> internal/cli/cli.go
