+++ cmd/binq/main.go
@@ -3,9 +3,9 @@ package main
 import (
 	"os"
 
-	"github.com/progrhyme/binq/client"
+	"github.com/progrhyme/binq/internal/cli"
 )
 
 func main() {
-	os.Exit(client.NewCLI(os.Stdout, os.Stderr).Run(os.Args))
+	os.Exit(cli.NewCLI(os.Stdout, os.Stderr).Run(os.Args))
 }
